Change the cpp color type for uint8_t for now

+++ api/cpp/cbindgen.rs
@@ -234,7 +234,7 @@ fn default_config() -> cbindgen::Config {
             // therefore it is ok to reinterpret_cast
             ("MenuEntryModel".into(), "std::shared_ptr<slint::Model<MenuEntry>>".into()),
             ("Coord".into(), "float".into()),
-            ("Channel".into(), "float".into()),
+            ("Channel".into(), "uint8_t".into()),
         ]
         .iter()
         .cloned()
